added get all products

+++ backend/shop/migrations/000001_products.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE products (
                           id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
                           name TEXT NOT NULL,
                           description TEXT,
-                          price NUMERIC(10, 2) NOT NULL,
+                          price int NOT NULL,
                           category TEXT,
                           image_url TEXT,
                           created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
